feat: complete macro consolidation (issue #201)
- Migrated all macros to new src/macros/ module structure
- Clear separation between public API and internal macros
- Removed namespace pollution at crate root
- Updated all import paths throughout codebase
- Added comprehensive documentation
- All tests passing, clippy clean

Breaking changes: None - public API unchanged
