+++ src/command/nd_filter.rs
@@ -10,10 +10,9 @@
 //! - All ND filter commands - Sony FR7 specific
 //! - The FR7 supports variable ND filter (2 to 7 stops, continuously variable)
 
-use crate::{
-    command::const_encoding::constants, error::Error, visca_bool_command, visca_builder,
-    visca_param_command,
-};
+use crate::macros::internal::*;
+
+use crate::{command::const_encoding::constants, error::Error};
 
 /// ND filter mode for Sony FR7.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -168,7 +167,7 @@ visca_bool_command! {
 #[allow(clippy::expect_used)]
 mod tests {
     use super::*;
-    use crate::visca_test;
+    use crate::macros::test_utils::visca_test;
 
     visca_test!(
         NDFilterModeCommand,
